Deleted references to methods being called in views.
