+++ app/views/wrestlers/new.html.erb
@@ -1,4 +1,3 @@
-<% initial_stats(@wrestler) %>
 <h1 class="title">Make a New Wrestler From Scratch</h1>
 <div class="wrestler-display-container">
 	<%= render "shared/form", wrestler: @wrestler %>
