+++ app/views/wrestlers/show.html.erb
@@ -1,4 +1,3 @@
-<% generate_wrestler_stats(@wrestler) %>
 <h1 class="title">Current Wrestler Card Values & Stats</h1>
 
 <div class="title wrestler-download-wrapper">
